move the padding utility to a new file
